Add the documentation for the move module and move it to the primitive package.
